move related deps to proper components
