+++ package.json
@@ -31,7 +31,8 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "typescript": "4.9.5",
-    "wagmi": "^0.12.13"
+    "wagmi": "^0.12.13",
+    "zustand": "^4.3.8"
   },
   "devDependencies": {
     "@types/react-modal": "^3.16.0",
